+++ clamor/models/webhook.py
@@ -1,0 +1,5 @@
+from .base import Base
+
+
+class Webhook(Base):
+    pass
