--- clamor/rest/endpoints/__init__.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import base
-from .audit_log import *
-from .channel import *
-from .emoji import *
-from .gateway import *
-from .guild import *
-from .invite import *
-from .oauth import *
-from .user import *
-from .voice import *
-from .webhook import *
